Updated logic for holidays and weekends
Implement real logic to check if any weekends fall within the rental range and track how many days within the weekend.
Implement real logic for holiday checks. Labor day is checked by seeing if the day is monday and within the first 7 days of the calendar, if these are true then it is the first monday of Sept ie Labor day.